huawei-solar-modbus-mqtt/requirements.txt
# Core Dependencies
huawei-solar>=2.5.0
pymodbus>=3.11.4

# MQTT Support
paho-mqtt>=2.1.0
